+++ src/main/java/pl/asie/computronics/integration/cofh/DriverEnergyProvider.java
@@ -1,6 +1,7 @@
 package pl.asie.computronics.integration.cofh;
 
-import cofh.api.energy.IEnergyHandler;
+import cofh.api.energy.IEnergyProvider;
+import cofh.api.energy.IEnergyReceiver;
 import dan200.computercraft.api.lua.ILuaContext;
 import dan200.computercraft.api.lua.LuaException;
 import dan200.computercraft.api.peripheral.IComputerAccess;
@@ -13,15 +14,15 @@ import pl.asie.computronics.reference.Names;
 /**
  * @author Vexatos
  */
-public class DriverEnergyHandler {
+public class DriverEnergyProvider {
 
-	public static class CCDriver extends CCMultiPeripheral<IEnergyHandler> {
+	public static class CCDriver extends CCMultiPeripheral<IEnergyProvider> {
 
 		public CCDriver() {
 		}
 
-		public CCDriver(IEnergyHandler tile, World world, int x, int y, int z) {
-			super(tile, Names.EnderIO_PoweredTile, world, x, y, z);
+		public CCDriver(IEnergyProvider tile, World world, int x, int y, int z) {
+			super(tile, Names.CoFH_PoweredTile, world, x, y, z);
 		}
 
 		@Override
@@ -32,8 +33,8 @@ public class DriverEnergyHandler {
 		@Override
 		public CCMultiPeripheral getPeripheral(World world, int x, int y, int z, int side) {
 			TileEntity te = world.getTileEntity(x, y, z);
-			if(te != null && te instanceof IEnergyHandler) {
-				return new CCDriver((IEnergyHandler) te, world, x, y, z);
+			if(te != null && te instanceof IEnergyProvider && !(te instanceof IEnergyReceiver)) {
+				return new CCDriver((IEnergyProvider) te, world, x, y, z);
 			}
 			return null;
 		}
@@ -48,12 +49,12 @@ public class DriverEnergyHandler {
 			if(arguments.length > 0 && !(arguments[0] instanceof Integer)) {
 				throw new LuaException("first argument needs to be a number or nil");
 			}
-			switch(method){
-				case 0:{
+			switch(method) {
+				case 0: {
 					final ForgeDirection side = arguments.length > 0 ? ForgeDirection.getOrientation((Integer) arguments[0]) : ForgeDirection.UNKNOWN;
 					return new Object[] { tile.getEnergyStored(side) };
 				}
-				case 1:{
+				case 1: {
 					final ForgeDirection side = arguments.length > 0 ? ForgeDirection.getOrientation((Integer) arguments[0]) : ForgeDirection.UNKNOWN;
 					return new Object[] { tile.getMaxEnergyStored(side) };
 				}