+++ src/main/java/pl/asie/computronics/item/ItemTape.java
@@ -34,10 +34,10 @@ import java.util.List;
 })
 public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaProvider {
 	public static final int L_SECOND = 4096;
-	public static final int L_MINUTE = 4096*60;
+	public static final int L_MINUTE = 4096 * 60;
 
 	private static final int TAPE_COUNT = 10;
-	private static final int[] DEFAULT_LENGTHS = { 4, 8, 16, 32, 64, 2, 6, 16, 128, 128};
+	private static final int[] DEFAULT_LENGTHS = { 4, 8, 16, 32, 64, 2, 6, 16, 128, 128 };
 
 	private int[] sizes;
 	private IIcon tape_i, tape_g, tape_d, tape_n, tape_c, tape_co, tape_st, tape_greg, tape_ig;
@@ -53,14 +53,35 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 
 		// parse lengths
 		String[] l = lengths.split(",");
-		sizes = new int[l.length];
+
+		/*if(l.length < TAPE_COUNT) {
+			Computronics.log.error("Config error: Property 'tapedrive.tapeLengths' in computronics.cfg"
+				+ " must contain 10 numbers separated by commas. Setting values to default"
+				+ " [default: 4,8,16,32,64,2,6,16,128,128]");
+			Config.TAPE_LENGTHS = "4,8,16,32,64,2,6,16,128,128";
+			lengths = Config.TAPE_LENGTHS;
+			l = lengths.split(",");
+		}*/
+
+		sizes = new int[TAPE_COUNT];
 		for(int i = 0; i < TAPE_COUNT; i++) {
-			if(i<l.length) {
-				sizes[i] = Integer.parseInt(l[i]) * L_MINUTE;
-				if(sizes[i] <= 0)
+			if(i < l.length) {
+				try {
+					sizes[i] = Integer.parseInt(l[i]) * L_MINUTE;
+				} catch(NumberFormatException e) {
+					Computronics.log.error("Property 'tapedrive.tapeLengths' in computronics.cfg contains entry"
+						+ " that is not a number! Setting entry "
+						+ String.valueOf(i + 1) + " ['" + l[i] + "'] to " + DEFAULT_LENGTHS[i]);
+					e.printStackTrace();
+					sizes[i] = DEFAULT_LENGTHS[i] * L_MINUTE;
+				}
+				if(sizes[i] <= 0) {
 					sizes[i] = 4;
+				}
 			} else {
-				sizes[i]=DEFAULT_LENGTHS[i] * L_MINUTE;
+				Computronics.log.warn("Property 'tapedrive.tapeLengths' contains too few entries, setting entry "
+					+ String.valueOf(i + 1) + " to " + DEFAULT_LENGTHS[i]);
+				sizes[i] = DEFAULT_LENGTHS[i] * L_MINUTE;
 			}
 		}
 	}
@@ -81,21 +102,35 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 
 	@Override
 	public IIcon getIconFromDamageForRenderPass(int meta, int pass) {
-		if(pass == 0) switch(meta) {
-			case 0: return tape_i;
-			case 1: return tape_g;
-			case 2: return tape_g;
-			case 3: return tape_d;
-			case 4: return tape_n;
-			case 5: return tape_co;
-			case 6: return tape_st;
-			case 7: return tape_greg;
-			case 8: return tape_n;
-			case 9: return tape_ig;
+		if(pass == 0) {
+			switch(meta) {
+				case 0:
+					return tape_i;
+				case 1:
+					return tape_g;
+				case 2:
+					return tape_g;
+				case 3:
+					return tape_d;
+				case 4:
+					return tape_n;
+				case 5:
+					return tape_co;
+				case 6:
+					return tape_st;
+				case 7:
+					return tape_greg;
+				case 8:
+					return tape_n;
+				case 9:
+					return tape_ig;
 
-			default: return tape_i;
+				default:
+					return tape_i;
+			}
+		} else {
+			return tape_c;
 		}
-		else return tape_c;
 	}
 
 	public String getLabel(ItemStack stack) {
@@ -104,7 +139,9 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 	}
 
 	public boolean setLabel(ItemStack stack, String label) {
-		if(stack == null) return false;
+		if(stack == null) {
+			return false;
+		}
 		stack.getTagCompound().setString("label", label);
 		return true;
 	}
@@ -113,19 +150,21 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 	@Override
 	public void addInformation(ItemStack stack, EntityPlayer player, List text, boolean par4) {
 		int size = getSize(stack);
-		int len = (int)Math.floor(size / L_MINUTE);
+		int len = (int) Math.floor(size / L_MINUTE);
 		if(stack.getTagCompound() != null) {
 			String label = getLabel(stack);
-			if(label.length() > 0) text.add(EnumChatFormatting.WHITE + "" + EnumChatFormatting.ITALIC + label);
+			if(label.length() > 0) {
+				text.add(EnumChatFormatting.WHITE + "" + EnumChatFormatting.ITALIC + label);
+			}
 		}
-		text.add(EnumChatFormatting.GRAY + StringUtil.localizeAndFormat("tooltip.computronics.tape.length", ""+len));
+		text.add(EnumChatFormatting.GRAY + StringUtil.localizeAndFormat("tooltip.computronics.tape.length", "" + len));
 
-		switch (stack.getItemDamage()){
-			case 7:{
+		switch(stack.getItemDamage()) {
+			case 7: {
 				text.add(EnumChatFormatting.AQUA + StringUtil.localize("tooltip.computronics.tape.balanced"));
 				break;
 			}
-			case 9:{
+			case 9: {
 				String[] local = StringUtil.localize("tooltip.computronics.tape.ig")
 					.replace("\\n", "\n").split("\\n");
 				for(String s : local) {
@@ -133,14 +172,15 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 				}
 				break;
 			}
-			default: break;
+			default:
+				break;
 		}
 	}
 
 	@SuppressWarnings("unchecked")
 	@Override
 	@SideOnly(Side.CLIENT)
-    public void getSubItems(Item item, CreativeTabs tabs, List list) {
+	public void getSubItems(Item item, CreativeTabs tabs, List list) {
 		for(int i = 0; i < TAPE_COUNT; i++) {
 			if((i == 7 || i == 9) && !Loader.isModLoaded(Mods.GregTech)) {
 				//Do nothing. If we return here, we lose all new tapes.
@@ -148,11 +188,12 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 			}
 			list.add(new ItemStack(item, 1, i));
 		}
-     }
+	}
 
 	public int getSize(ItemStack stack) {
 		return getSize(stack.getItemDamage());
 	}
+
 	public int getSize(int meta) {
 		return sizes[meta % sizes.length];
 	}
@@ -164,21 +205,25 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 			// Exists, read NBT data if everything is alright
 			NBTTagCompound nbt = stack.getTagCompound();
 			String storageName = nbt.getString("storage");
-			if(Computronics.storage.exists(storageName))
+			if(Computronics.storage.exists(storageName)) {
 				return Computronics.storage.get(storageName, size, 0);
+			}
 		}
 
 		// Doesn't exist, create new storage and write NBT data
 		TapeStorage storage = Computronics.storage.newStorage(size);
-		if(stack.getTagCompound() == null) stack.setTagCompound(new NBTTagCompound());
+		if(stack.getTagCompound() == null) {
+			stack.setTagCompound(new NBTTagCompound());
+		}
 		stack.getTagCompound().setString("storage", storage.getUniqueId());
 		return storage;
 	}
 
 	@Override
 	public String getUnlocalizedName(ItemStack itemstack) {
-		if(itemstack != null && itemstack.getItemDamage() == 9)
+		if(itemstack != null && itemstack.getItemDamage() == 9) {
 			return "item.computronics.tape.ig";
+		}
 		return super.getUnlocalizedName(itemstack);
 	}
 
@@ -191,7 +236,7 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 	public Entity createEntity(World world, Entity location, ItemStack itemstack) {
 		if(itemstack != null && itemstack.getItemDamage() == 9) {
 			EntityItemIndestructable newTapeEntity = new EntityItemIndestructable(
-				world,location.posX,location.posY,location.posZ, itemstack);
+				world, location.posX, location.posY, location.posZ, itemstack);
 			newTapeEntity.delayBeforeCanPickup = 40;
 			newTapeEntity.motionX = location.motionX;
 			newTapeEntity.motionY = location.motionY;
@@ -208,26 +253,35 @@ public class ItemTape extends Item implements IItemTapeStorage, IMedia, IMediaPr
 	}
 
 	@Override
-    @SideOnly(Side.CLIENT)
-    public int getColorFromItemStack(ItemStack stack, int pass)
-    {
-        return pass == 0 ? 16777215 : (ItemColorizer.hasColor(stack) ? ItemColorizer.getColor(stack) : 16777215);
-    }
+	@SideOnly(Side.CLIENT)
+	public int getColorFromItemStack(ItemStack stack, int pass) {
+		return pass == 0 ? 16777215 : (ItemColorizer.hasColor(stack) ? ItemColorizer.getColor(stack) : 16777215);
+	}
 
 	@Override
 	@Optional.Method(modid = Mods.ComputerCraft)
 	public IMedia getMedia(ItemStack stack) {
-		if(stack != null && stack.stackSize > 0 && stack.getItem() != null && stack.getItem() instanceof ItemTape) return ((IMedia)stack.getItem());
+		if(stack != null && stack.stackSize > 0 && stack.getItem() != null && stack.getItem() instanceof ItemTape) {
+			return ((IMedia) stack.getItem());
+		}
 		return null;
 	}
 
 	@Override
 	@Optional.Method(modid = Mods.ComputerCraft)
-	public String getAudioTitle(ItemStack stack) { return null; }
+	public String getAudioTitle(ItemStack stack) {
+		return null;
+	}
+
 	@Override
 	@Optional.Method(modid = Mods.ComputerCraft)
-	public String getAudioRecordName(ItemStack stack) { return null; }
+	public String getAudioRecordName(ItemStack stack) {
+		return null;
+	}
+
 	@Override
 	@Optional.Method(modid = Mods.ComputerCraft)
-	public IMount createDataMount(ItemStack stack, World world) { return null; }
+	public IMount createDataMount(ItemStack stack, World world) {
+		return null;
+	}
 }