+++ src/main/java/pl/asie/computronics/network/NetworkHandlerClient.java
@@ -7,6 +7,7 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.network.INetHandler;
 import net.minecraft.tileentity.TileEntity;
 import pl.asie.computronics.Computronics;
+import pl.asie.computronics.oc.DriverCardBoom;
 import pl.asie.computronics.oc.DriverCardSound;
 import pl.asie.computronics.reference.Config;
 import pl.asie.computronics.reference.Mods;
@@ -25,17 +26,18 @@ public class NetworkHandlerClient extends MessageHandlerBase {
 
 	@Override
 	public void onMessage(Packet packet, INetHandler handler, EntityPlayer player, int command)
-			throws IOException {
+		throws IOException {
 		//System.out.println("CLIENT PACKET " + command);
 		switch(command) {
 			case Packets.PACKET_TAPE_GUI_STATE: {
 				TileEntity entity = packet.readTileEntity();
 				State state = State.values()[packet.readUnsignedByte()];
 				if(entity instanceof TileTapeDrive) {
-					TileTapeDrive tile = (TileTapeDrive)entity;
+					TileTapeDrive tile = (TileTapeDrive) entity;
 					tile.switchState(state);
 				}
-			} break;
+			}
+			break;
 			case Packets.PACKET_AUDIO_DATA: {
 				int dimId = packet.readInt();
 				int x = packet.readInt();
@@ -47,15 +49,17 @@ public class NetworkHandlerClient extends MessageHandlerBase {
 				short volume = packet.readByte();
 				byte[] data = packet.readByteArrayData(packetSize);
 				byte[] audio = new byte[packetSize * 8];
-				String sourceName = "dfpwm_"+codecId;
+				String sourceName = "dfpwm_" + codecId;
 				StreamingAudioPlayer codec = Computronics.instance.audio.getPlayer(codecId);
 
-				if(dimId != WorldUtils.getCurrentClientDimension()) return;
+				if(dimId != WorldUtils.getCurrentClientDimension()) {
+					return;
+				}
 
 				codec.decompress(audio, data, 0, 0, packetSize);
 				for(int i = 0; i < (packetSize * 8); i++) {
 					// Convert signed to unsigned data
-					audio[i] = (byte)(((int)audio[i] & 0xFF) ^ 0x80);
+					audio[i] = (byte) (((int) audio[i] & 0xFF) ^ 0x80);
 				}
 
 				if((codec.lastPacketId + 1) != packetId) {
@@ -63,30 +67,40 @@ public class NetworkHandlerClient extends MessageHandlerBase {
 				}
 				codec.setSampleRate(packetSize * 32);
 				codec.setDistance((float) Config.TAPEDRIVE_DISTANCE);
-				codec.setVolume(volume/127.0F);
+				codec.setVolume(volume / 127.0F);
 				codec.playPacket(audio, x, y, z);
 				codec.lastPacketId = packetId;
-			} break;
+			}
+			break;
 			case Packets.PACKET_AUDIO_STOP: {
 				int codecId = packet.readInt();
 				Computronics.instance.audio.removePlayer(codecId);
-			} break;
+			}
+			break;
 			case Packets.PACKET_PARTICLE_SPAWN: {
-		        double x = packet.readFloat();
-		        double y = packet.readFloat();
-		        double z = packet.readFloat();
-		        double vx = packet.readFloat();
-		        double vy = packet.readFloat();
-		        double vz = packet.readFloat();
-		        String name = packet.readString();
-		        Minecraft.getMinecraft().thePlayer.getEntityWorld().spawnParticle(name, x, y, z, vx, vy, vz);
-			} break;
+				double x = packet.readFloat();
+				double y = packet.readFloat();
+				double z = packet.readFloat();
+				double vx = packet.readFloat();
+				double vy = packet.readFloat();
+				double vz = packet.readFloat();
+				String name = packet.readString();
+				Minecraft.getMinecraft().thePlayer.getEntityWorld().spawnParticle(name, x, y, z, vx, vy, vz);
+			}
+			break;
 			case Packets.PACKET_COMPUTER_BEEP: {
 				if(Loader.isModLoaded(Mods.OpenComputers)) {
 					DriverCardSound.onSound(packet, player);
 				}
-			} break;
-			case 5: {
+			}
+			break;
+			case Packets.PACKET_COMPUTER_BOOM: {
+				if(Loader.isModLoaded(Mods.OpenComputers)) {
+					Computronics.proxy.goBoom(packet);
+				}
+			}
+			break;
+			case 6: {
 				if(Mods.API.hasClass("marytts.LocalMaryInterface")){
 					Computronics.tts.say(packet.readString());
 				}