+++ src/main/java/pl/asie/computronics/oc/IntegrationOpenComputers.java
@@ -6,6 +6,7 @@ import cpw.mods.fml.common.event.FMLMissingMappingsEvent;
 import cpw.mods.fml.common.registry.GameRegistry;
 import li.cil.oc.api.Driver;
 import net.minecraft.block.Block;
+import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
 import org.apache.logging.log4j.Logger;
@@ -15,6 +16,7 @@ import pl.asie.computronics.integration.betterstorage.DriverCrateStorageNew;
 import pl.asie.computronics.integration.betterstorage.DriverCrateStorageOld;
 import pl.asie.computronics.integration.buildcraft.DriverHeatable;
 import pl.asie.computronics.integration.buildcraft.pluggable.IntegrationBuildCraft;
+import pl.asie.computronics.integration.draconicevolution.DriverExtendedRFStorage;
 import pl.asie.computronics.integration.enderio.DriverAbstractMachine;
 import pl.asie.computronics.integration.enderio.DriverAbstractPoweredMachine;
 import pl.asie.computronics.integration.enderio.DriverCapacitorBank;
@@ -72,7 +74,11 @@ public class IntegrationOpenComputers {
 	@Optional.Method(modid = Mods.OpenComputers)
 	public void preInit() {
 
-		if(Config.OC_ROBOT_UPGRADES || Config.OC_CARD_FX || Config.OC_CARD_SPOOF || Config.OC_CARD_SOUND) {
+		if(Config.OC_ROBOT_UPGRADES
+			|| Config.OC_CARD_FX
+			|| Config.OC_CARD_SPOOF
+			|| Config.OC_CARD_SOUND
+			|| Config.OC_CARD_BOOM) {
 			itemOCParts = new ItemOpenComputers();
 			GameRegistry.registerItem(itemOCParts, "computronics.ocParts");
 			Driver.add(itemOCParts);
@@ -175,6 +181,11 @@ public class IntegrationOpenComputers {
 			}
 		}
 
+		if(Loader.isModLoaded(Mods.API.DraconicEvolution)
+			&& compat.isCompatEnabled(Compat.DraconicEvolution)) {
+			Driver.add(new DriverExtendedRFStorage.OCDriver());
+		}
+
 		if(Mods.API.hasVersion(Mods.API.BuildCraftTiles, "[1.1,)")) {
 			if(compat.isCompatEnabled(Compat.BuildCraft_Drivers)) {
 				Driver.add(new DriverHeatable.OCDriver());
@@ -235,6 +246,14 @@ public class IntegrationOpenComputers {
 				'f', Computronics.ironNote,
 				'b', li.cil.oc.api.Items.get("card").createItemStack(1),
 				'l', li.cil.oc.api.Items.get("cu").createItemStack(1));
+		}
+		if(Config.OC_CARD_BOOM) {
+			GameRegistry.addShapedRecipe(new ItemStack(itemOCParts, 1, 6),
+				"mf", "fb",
+				'm', li.cil.oc.api.Items.get("cu").createItemStack(1),
+				'f', Blocks.tnt,
+				'b', li.cil.oc.api.Items.get("redstoneCard1").createItemStack(1));
+
 		}
 		if(Computronics.buildcraft != null) {
 			Computronics.buildcraft.postInitOC();